Revert "Aula 68"
This reverts commit 2af11f2a6756503ef031515d27649ebe8f2df5b0.

+++ src/main/java/domain/CarrosService.java
@@ -8,7 +8,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
-//Serviço
+import domain.Carro;
+import domain.CarroDTO;
+import domain.CarrosRepository;
+
 @Service
 public class CarrosService {
 	@Autowired
@@ -24,12 +27,11 @@ public class CarrosService {
 		return list;
 	}
 	
-	  public CarroDTO getCarroById(Long id) {
-	    Optional<Carro> carro = rep.findById(id);
-        return carro.map(CarroDTO::create).orElseThrow(() -> new domain.ObjectNotFoundException("Carro não encontrado"));
+	public Optional<Carro> getCarrosbyId(Long id) {
+		return rep.findById(id);
+		
 	}
 	
-
 	public List<CarroDTO> getCarrosbyTipo(String tipo) {
 		List<Carro> carros  = rep.findByTipo(tipo);
 		List<CarroDTO> list  = new ArrayList();		
@@ -40,17 +42,17 @@ public class CarrosService {
 		return list;		
 	}
 	
-	public CarroDTO insert(Carro carro) {
+	public Carro insert(Carro carro) {
 		Assert.isNull((carro.getId()), "Não foi possivel inserir o registro");
-		return CarroDTO.create(rep.save(carro));
+		return rep.save(carro);
 		
 	}	
 	
-	public CarroDTO update(Carro carro, Long id) {
+	public Carro update(Carro carro, Long id) {
 		Assert.notNull(id, "Não foi possivel atualizar o registro");
 		
 		//Busca Carro no banco de dados
-		Optional <Carro> optional = rep.findById(id);
+		Optional <Carro> optional = getCarrosbyId(id);
 		if (optional.isPresent()) {
 		//Copiar as propriedades
 		Carro db = optional.get();
@@ -61,14 +63,28 @@ public class CarrosService {
 		//Atualizar o carro
 		rep.save(db);
 		
-		return CarroDTO.create(db);
+		return db;
 		}else{
-		
 		 throw new RuntimeException("Não foi possivel  atualizar o registro");
 		}
 	}	
 	
-	public void delete(Long id) {
+	public boolean delete(Long id) {
+		Optional <Carro> carro = getCarrosbyId(id);
+		if(carro.isPresent()) {
 	      rep.deleteById(id);
+		  return true;
+		}
+		  return false;
 	}
+	
+//   public List<Carro> getCarrosFake() {
+//		List<Carro> carros = new ArrayList<>();
+//		
+//		carros.add(new Carro(1L, "Fusca", "esportivo"));
+//		carros.add(new Carro(2L, "Gol", "esportivo"));
+//		carros.add(new Carro(3L, "Brasilia", "esportivo"));
+//
+//		return carros;
+//   }
 }
